feat(project): moved data sets to utils, improved xpense authorization
