+++ frontends/wrap-r/src/services/github-auth-provider.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@furystack/inject'
 import { WrapRApiService, SessionService } from 'common-frontend-utils'
-import { tokens } from 'common-config/src'
+import { tokens } from 'common-config'
 
 @Injectable({ lifetime: 'singleton' })
 export class GithubAuthProvider {
